refactor. searchbar 컴포넌트 리팩토링(search bar layer 에서 하위 컴포넌트 로직 모두 일괄처리할 수 있도록 파일 구조 수정)
