--- src/layer/SearchInputLayer.tsx
@@ -1,19 +1,0 @@
-import { VAC, VACInput } from 'react-vac';
-
-import SearchInput from 'components/atoms/SearchInput/SearchInput';
-
-const SearchInputLayer = ({ value, onChange }: any) => {
-  const searchInputViewProps = {
-    value,
-    onChange,
-  };
-
-  return (
-    <>
-      <SearchInput {...searchInputViewProps} />
-      {/* <VACInput name="SearchInput" data={searchInputViewProps} /> */}
-    </>
-  );
-};
-
-export default SearchInputLayer;
